Extract importer adaptive tuning module
