+++ tests/Import/AdaptiveTunerTest.php
@@ -2,10 +2,10 @@
 
 namespace ImportTests;
 
-use AdaptiveTuner;
 use PHPUnit\Framework\TestCase;
+use Reprint\Importer\Tuning\AdaptiveTuner;
 
-require_once __DIR__ . '/../../importer/import.php';
+require_once __DIR__ . '/../../packages/reprint-importer/src/lib/tuning/class-adaptive-tuner.php';
 
 class AdaptiveTunerTest extends TestCase
 {
